requirements.txt
tldextract
scipy
nltk
appdirs>=1.4.3
asn1crypto>=1.3.0
certifi>=2019.11.28
cffi>=1.13.2
chardet>=3.0.4
cryptography>=2.8
docker>=4.2.0
docker-pycreds>=0.4.0
idna>=2.9
numpy>=1.18.1
packaging>=20.1
pandas>=1.0.1
pycparser>=2.19
pyOpenSSL>=19.1.0
pyparsing>=2.4.6
PySocks>=1.7.1
python-dateutil>=2.8.1
pytz>=2019.3
requests>=2.23.0
selenium>=3.141.0
six>=1.14.0
urllib3>=1.25.7
websocket-client>=0.57.0
chromedriver-binary>=2.38

